got everything into a working start, created league blueprint
