+++ src/app/league/__init__.py
@@ -1,3 +1,1 @@
 __author__ = 'djw'
-
-
