+++ src/app/league/models.py
@@ -1,5 +1,6 @@
-from .base import BaseModel
+
 from google.appengine.ext import ndb
+from app.models import BaseModel
 import tinyid
