+++ src/app/models.py
@@ -1,5 +1,7 @@
 from google.appengine.ext import ndb
 
+__author__ = 'djw'
+
 
 class BaseModel(ndb.Model):
     KEY_NAME_FIELDS = []
