+++ src/app/user/views.py
@@ -4,7 +4,7 @@ from flask import Blueprint, redirect, request, g, url_for, render_template
 from .forms import SettingsForm
 from .models import update_user, User, create_user
 
-user = Blueprint('user', __name__, url_prefix='/user')
+user_module = Blueprint('user', __name__, url_prefix='/user')
 
 
 def login_required(func):
@@ -29,7 +29,7 @@ def get_authed_user():
     return None
 
 
-@user.route('/settings/', methods=['GET', 'POST'])
+@user_module.route('/settings/', methods=['GET', 'POST'])
 @login_required
 def settings():
     form = SettingsForm(request.form, obj=g.user)
@@ -40,7 +40,7 @@ def settings():
     return render_template('user/settings.html', form=form, **g.context)
 
 
-@user.route('/dashboard/')
+@user_module.route('/dashboard/')
 @login_required
 def user_dashboard():
     if not g.is_logged_in:
