+++ src/templates/league/create_league.html
@@ -2,8 +2,8 @@
 {% from "_formhelpers.html" import render_field %}
 {% block content %}
 <div class="container col-md-6">
-<h2>Edit League</h2>
-<form method=post action={{ url_for('edit_league_form',league_id=league.league_id) }}>
+<h2>Create League</h2>
+<form method=post action="{{ url_for('league.create_league_form') }}">
   {{ render_field(form.name) }}
   {{ render_field(form.rating_scheme) }}
   {{ render_field(form.description) }}
