+++ src/templates/league/edit_league.html
@@ -2,8 +2,8 @@
 {% from "_formhelpers.html" import render_field %}
 {% block content %}
 <div class="container col-md-6">
-<h2>Create League</h2>
-<form method=post action="{{ url_for('create_league_form') }}">
+<h2>Edit League</h2>
+<form method=post action={{ url_for('league.edit_league_form',league_id=league.league_id) }}>
   {{ render_field(form.name) }}
   {{ render_field(form.rating_scheme) }}
   {{ render_field(form.description) }}
