+++ src/templates/menu.html
@@ -15,7 +15,7 @@
             <ul class="nav navbar-nav">
                 <li><a href="{{ url_for('user.user_dashboard') }}">Dashboard</a></li>
                 <li><a href="{{ url_for('user.settings') }}">Settings</a></li>
-                <li><a href="{{ url_for('display_leagues') }}">Leagues</a></li>
+                <li><a href="{{ url_for('league.display_leagues') }}">Leagues</a></li>
                 <li><a href="#contact">Support</a></li>
             </ul>
             {% endif %}
